feat: 전 회차 정보 컴포넌트로 작업
