+++ src/components/latest/style.ts
@@ -1,0 +1,20 @@
+import styled from "styled-components";
+
+export const StyledContainer = styled.div`
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  margin-top: 20px;
+`;
+
+export const StyledNumbersContainer = styled.div`
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  gap: 10px;
+`;
+
+export const DrawHighlight = styled.span`
+  color: #FF6868;
+  font-weight: bold;
+`;
